Update main-custom.js
Cleaning up most of the testing items that were commented out.  Now have two identical functions that -- if we need to iterate through some designs or if we can turn it on in an unused collection to have folks see it live... to turn on for Nash collection just need to change the collection scope.